Move interp into utils
